+++ src/core/conversion/mod.rs
@@ -1,6 +1,7 @@
 
 pub mod scalar_float_to_integer;
 pub mod vector_float_to_integer;
+pub mod scalar_float;
 pub mod scalar_to_string;
 pub mod vector_to_string;
 pub mod scalar_to_byte;